Change BasicConnection to do "pooling"
All connections in a single thread will share a single zmq socket.  This
could be further optimized in the future to maintain a map of queues,
but for now we'll assume that if the number of threads you run is larger
than the number of file descriptors available, you have other problems.